Re-enable validation options (:if, :unless, etc).
Previous to this commit, we passed :on => :save as the default option to
the validation. Confusingly, this is not a valid option for the :on
parameter. :on only will take one of:

- :create
- :update

Otherwise, it infers whether to set :on to :create or :update based on
the record's state. Something like:

  options[:on] = new_record? ? :create : :update

This also adds full test coverage for this issue.

Closes issue #9.