allow :field as valid option key, fixes #27
The Mongoid integration depends on the :field key to set additional
options for automatically generated field. When setting a default
option for :field it fails, because as_enum asserts the option keys.

Implemented fix just allows :field as valid option key for as_enum,
eventhough it is a no-op (except for mongoid).